ai_agents/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.8 